Make vim loading plugins and keymaps more dynamic
nvim will iterate through all the available files in plugins and keymap
at startup to load them, as available.

The setup.sh script prompts before symlinking any non-lua files from
keymap and plugins directories.

+++ nvim/setup.sh
@@ -1,30 +1,58 @@
 #!/usr/bin/env bash
+config_dir=~/.config/nvim
 
 # Backup existing config
 cp -r ~/.config/nvim ~/.config/nvim.bak
 
-mkdir ~/.config/nvim
-mkdir ~/.config/nvim/lua
+# Recursively create directories if they don't exist
+mkdir -p "$config_dir/lua"
+
+# Prompt the user before symlinking the init.lua file
+echo "Symlink nvim/init.lua? (y/n)"
+read -r response
+if [ "$response" = "y" ]; then
+  rm -rf "$config_dir/init.lua"
+  ln -s $(pwd)/init.lua "$config_dir/init.lua"
+fi
 
 # Symlink individual files and directories under lua
-rm  -rf ~/.config/nvim/lua/color
-ln -s $(pwd)/lua/color ~/.config/nvim/lua/color
+rm  -rf "$config_dir/lua/color"
+ln -s $(pwd)/lua/color "$config_dir/lua/color"
 
-rm -rf ~/.config/nvim/lua/copilot
-ln -s $(pwd)/lua/copilot ~/.config/nvim/lua/copilot
+rm  -rf "$config_dir/lua/options"
+ln -s $(pwd)/lua/options "$config_dir/lua/options"
 
-rm -rf ~/.config/nvim/lua/options
-ln -s $(pwd)/lua/options ~/.config/nvim/lua/options
+# Symlink files inside the keymap directory. Prompt the user for each file
+# before symlinking it, with the exception of init.lua
+mkdir -p "$config_dir/lua/keymap"
+for file in $(pwd)/lua/keymap/*; do
+  if [ "$file" = "$(pwd)/lua/keymap/init.lua" ]; then
+    rm -rf "$config_dir/lua/keymap/$(basename $file)"
+    ln -s $file "$config_dir/lua/keymap/$(basename $file)"
+    continue
+  fi
+  echo "Symlink $file? (y/n)"
+  read -r response
+  if [ "$response" = "y" ]; then
+    rm -rf "$config_dir/lua/keymap/$(basename $file)"
+    ln -s $file "$config_dir/lua/keymap/$(basename $file)"
+  fi
+done
 
-rm -rf ~/.config/nvim/lua/keymap.lua
-ln -s $(pwd)/lua/keymap.lua ~/.config/nvim/lua/keymap.lua
-
-rm -rf ~/.config/nvim/lua/keymap.vim
-ln -s $(pwd)/lua/keymap.vim ~/.config/nvim/lua/keymap.vim
-
+# plugins that are loaded dynamically. 
+# We only symlink the init.lua file and prompt the user before symlink-ing
+# any of any of the other ones.
 mkdir ~/.config/nvim/lua/plugins
-rm -rf ~/.config/nvim/lua/plugins/init.lua
-ln -s $(pwd)/lua/plugins/init.lua ~/.config/nvim/lua/plugins/init.lua
+for file in $(pwd)/lua/plugins/*; do
+  if [ "$file" = "$(pwd)/lua/plugins/init.lua" ]; then
+    rm -rf ~/.config/nvim/lua/plugins/init.lua
+    ln -s $file ~/.config/nvim/lua/plugins/$(basename $file)
+    continue
+  fi
+  echo "Symlink $file? (y/n)"
+  read -r response
+  if [ "$response" = "y" ]; then
+    ln -s $file ~/.config/nvim/lua/plugins/$(basename $file)
+  fi
+done
 
-# Just copy the init.lua instead of symlink-ing it so that it can be customized for local usage
-# cp $(pwd)/init.lua ~/.config/nvim/init.lua
